Rename classes to modules
