+++ corebehrt/functional/filter.py
@@ -3,7 +3,7 @@ from typing import List
 
 import pandas as pd
 
-from corebehrt.classes.dataset import PatientData
+from corebehrt.modules.dataset import PatientData
 
 
 def exclude_short_sequences(
