+++ corebehrt/main/pretrain.py
@@ -2,8 +2,6 @@
 
 import logging
 
-from corebehrt.classes.dataset import MLMDataset
-from corebehrt.classes.prepare_data import DatasetPreparer
 from corebehrt.common.config import load_config
 from corebehrt.common.initialize import Initializer, ModelManager
 from corebehrt.common.loader import (
@@ -15,6 +13,8 @@ from corebehrt.functional.load import load_vocabulary
 from corebehrt.functional.save import save_pids_splits
 from corebehrt.functional.split import load_train_val_split, split_pids_into_train_val
 from corebehrt.functional.trainer_utils import replace_steps_with_epochs
+from corebehrt.modules.dataset import MLMDataset
+from corebehrt.modules.prepare_data import DatasetPreparer
 from corebehrt.trainer.trainer import EHRTrainer
 
 CONFIG_PATH = "./corebehrt/configs/pretrain.yaml"
