+++ corebehrt/modules/prepare_data.py
@@ -9,8 +9,8 @@ import pandas as pd
 import torch
 from dask.diagnostics import ProgressBar
 
-from corebehrt.classes.dataset import PatientDataset
-from corebehrt.classes.patient_filter import filter_df_by_pids
+from corebehrt.modules.dataset import PatientDataset
+from corebehrt.modules.patient_filter import filter_df_by_pids
 from corebehrt.common.config import Config, load_config
 from corebehrt.common.constants import ABSPOS_COL, PID_COL, TIMESTAMP_COL
 from corebehrt.common.setup import (
