+++ tests/test_classes/test_index_date_handler.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 import pandas as pd
 
-from corebehrt.classes.outcomes import IndexDateHandler
+from corebehrt.modules.outcomes import IndexDateHandler
 from corebehrt.common.constants import PID_COL, TIMESTAMP_COL
