+++ tests/test_functional/test_filters.py
@@ -2,7 +2,7 @@ import unittest
 import operator
 import pandas as pd
 
-from corebehrt.classes.dataset import PatientData
+from corebehrt.modules.dataset import PatientData
 from corebehrt.functional.filter import (
     exclude_short_sequences,
     censor_patient,
